+++ ChangeLog.md
@@ -11,6 +11,7 @@ The code was almost completely rewritten.
 * `--scc0` and `--no-scc-aware` disables the SCC-aware optimization
 * `--sd` option for semi-deterministic output (and not cut-deterministic)
 * `--no-reduction` is a synonym for `s0` (disables reductions by Spot)
+* `--highlight` colors states regarding the part of sDBA: violet for the 1st (nondeterministic) part, green for the 2nd (deterministic) part. Cut edges are colored by red.
 
 ### Changed
 * When a breakpoint is reached (R = B), track accepting transitions of the next level in B (it was set to ∅ before).